backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
trafilatura==2.0.0
google-genai==1.14.0
thefuzz==0.22.1
python-dotenv==1.1.0
pydantic==2.11.1
